got rid of weapon class--integrated into player class
